refactor: split dbal demo components
