+++ frontends/nextjs/src/components/misc/demos/dbal/index.ts
@@ -1,0 +1,4 @@
+export { BlobStorageDemo } from './BlobStorageDemo'
+export { CachedDataDemo } from './CachedDataDemo'
+export { DBALDemo } from './DBALDemo'
+export { KVStoreDemo } from './KVStoreDemo'
